requirements.txt
mysql.connector
flet
requests